fix(lint): split commands.rs under the 500-line gate, fix broken doc link, refresh prebuilt.html
Moves the JSON request-body helpers and the loopback HTTP request/response
cycle (tags_value, insert_opt, insert_json_opt, to_body, request,
print_body) out of commands.rs into a new sibling commands_http.rs. Pure
code move, no behavior change.

Also fixes a broken intra-doc link in command_path_resolution.rs
(`super::model::RoutineResponse` no longer resolves from its new location)
and rebuilds prebuilt.html, stale after this round's ui/ splits.
